requirements.txt
cloudpickle==2.2.1
plotly==5.21.0
python-dotenv==1.0.1
reflex==0.4.8.post1
